requirements.txt
bidict==0.22.1
blinker==1.7.0
CacheControl==0.13.1
cachetools==5.3.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==41.0.7
Deprecated==1.2.14
firebase-admin==6.3.0
Flask==3.0.0
Flask-SocketIO==5.3.6
gcloud==0.18.3
google-api-core==2.15.0
google-api-python-client==2.111.0
google-auth==2.25.2
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.14.0
google-cloud-storage==2.14.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.62.0
grpcio==1.60.0
grpcio-status==1.60.0
h11==0.14.0
httplib2==0.22.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.2
jwcrypto==1.5.0
MarkupSafe==2.1.3
msgpack==1.0.7
oauth2client==4.1.3
proto-plus==1.23.0
protobuf==4.25.1
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pycryptodome==3.19.0
PyJWT==2.8.0
pyparsing==3.1.1
Pyrebase4==4.7.1
python-engineio==4.8.0
python-jwt==4.1.0
python-socketio==5.10.0
requests==2.29.0
requests-toolbelt==0.10.1
rsa==4.9
simple-websocket==1.0.0
six==1.16.0
uritemplate==4.1.1
urllib3==1.26.18
Werkzeug==3.0.1
wrapt==1.16.0
wsproto==1.2.0
